// desktop/src/generated/synkronus-client/models/Observation.ts
export interface Observation {
    /**
     * 
     * @type {string}
     * @memberof Observation
     */
    observation_id: string;
    /**
     * 
     * @type {string}
     * @memberof Observation
     */
    form_type: string;
    /**
     * 
     * @type {string}
     * @memberof Observation
     */
    form_version: string;
    /**
     * Arbitrary JSON object containing form data
     * @type {object}
     * @memberof Observation
     */
    data: object;
    /**
     * 
     * @type {Date}
     * @memberof Observation
     */
    created_at: Date;
    /**
     * 
     * @type {Date}
     * @memberof Observation
     */
    updated_at: Date;
    /**
     * 
     * @type {Date}
     * @memberof Observation
     */
    synced_at?: Date | null;
    /**
     * 
     * @type {boolean}
     * @memberof Observation
     */
    deleted: boolean;
    /**
     * 
     * @type {ObservationGeolocation}
     * @memberof Observation
     */
    geolocation?: ObservationGeolocation | null;
    /**
     * Optional author/creator identifier for the observation (e.g. username)
     * @type {string}
     * @memberof Observation
     */
    author?: string | null;
    /**
     * Optional client device identifier for the observation
     * @type {string}
     * @memberof Observation
     */
    device_id?: string | null;
    /**
     * Optional list of string tags (labeling, extensions, data cleaning)
     * @type {Array<string>}
     * @memberof Observation
     */
    tags?: Array<string> | null;
}