// desktop/src/generated/synkronus-client/models/SyncPullRequest.ts
// @ts-nocheck
/* tslint:disable */
/* eslint-disable */
/**
 * Synkronus API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.6
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { SyncPullRequestSince } from './SyncPullRequestSince';
import {
    SyncPullRequestSinceFromJSON,
    SyncPullRequestSinceFromJSONTyped,
    SyncPullRequestSinceToJSON,
    SyncPullRequestSinceToJSONTyped,
} from './SyncPullRequestSince';

/**
 * 
 * @export
 * @interface SyncPullRequest
 */
export interface SyncPullRequest {
    /**
     * 
     * @type {string}
     * @memberof SyncPullRequest
     */
    client_id: string;
    /**
     * Optional body copy of epoch; header x-repository-generation wins when both are sent. Must match the server or the request returns 409.
     * @type {number}
     * @memberof SyncPullRequest
     */
    repository_generation?: number;
    /**
     * 
     * @type {SyncPullRequestSince}
     * @memberof SyncPullRequest
     */
    since?: SyncPullRequestSince;
    /**
     * 
     * @type {Array<string>}
     * @memberof SyncPullRequest
     */
    schema_types?: Array<string>;
}

/**
 * Check if a given object implements the SyncPullRequest interface.
 */
export function instanceOfSyncPullRequest(value: object): value is SyncPullRequest {
    if (!('client_id' in value) || value['client_id'] === undefined) return false;
    return true;
}

export function SyncPullRequestFromJSON(json: any): SyncPullRequest {
    return SyncPullRequestFromJSONTyped(json, false);
}

export function SyncPullRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): SyncPullRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'client_id': json['client_id'],
        'repository_generation': json['repository_generation'] == null ? undefined : json['repository_generation'],
        'since': json['since'] == null ? undefined : SyncPullRequestSinceFromJSON(json['since']),
        'schema_types': json['schema_types'] == null ? undefined : json['schema_types'],
    };
}

export function SyncPullRequestToJSON(json: any): SyncPullRequest {
    return SyncPullRequestToJSONTyped(json, false);
}

export function SyncPullRequestToJSONTyped(value?: SyncPullRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'client_id': value['client_id'],
        'repository_generation': value['repository_generation'],
        'since': SyncPullRequestSinceToJSON(value['since']),
        'schema_types': value['schema_types'],
    };
}