// desktop/src/generated/synkronus-client/models/SyncPushRequest.ts
// @ts-nocheck
/* tslint:disable */
/* eslint-disable */
/**
 * Synkronus API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.6
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Observation } from './Observation';
import {
    ObservationFromJSON,
    ObservationFromJSONTyped,
    ObservationToJSON,
    ObservationToJSONTyped,
} from './Observation';

/**
 * 
 * @export
 * @interface SyncPushRequest
 */
export interface SyncPushRequest {
    /**
     * 
     * @type {string}
     * @memberof SyncPushRequest
     */
    transmission_id: string;
    /**
     * 
     * @type {string}
     * @memberof SyncPushRequest
     */
    client_id: string;
    /**
     * Optional body copy of epoch; header x-repository-generation wins when both are sent.
     * @type {number}
     * @memberof SyncPushRequest
     */
    repository_generation?: number;
    /**
     * 
     * @type {Array<Observation>}
     * @memberof SyncPushRequest
     */
    records: Array<Observation>;
}

/**
 * Check if a given object implements the SyncPushRequest interface.
 */
export function instanceOfSyncPushRequest(value: object): value is SyncPushRequest {
    if (!('transmission_id' in value) || value['transmission_id'] === undefined) return false;
    if (!('client_id' in value) || value['client_id'] === undefined) return false;
    if (!('records' in value) || value['records'] === undefined) return false;
    return true;
}

export function SyncPushRequestFromJSON(json: any): SyncPushRequest {
    return SyncPushRequestFromJSONTyped(json, false);
}

export function SyncPushRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): SyncPushRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'transmission_id': json['transmission_id'],
        'client_id': json['client_id'],
        'repository_generation': json['repository_generation'] == null ? undefined : json['repository_generation'],
        'records': ((json['records'] as Array<any>).map(ObservationFromJSON)),
    };
}

export function SyncPushRequestToJSON(json: any): SyncPushRequest {
    return SyncPushRequestToJSONTyped(json, false);
}

export function SyncPushRequestToJSONTyped(value?: SyncPushRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'transmission_id': value['transmission_id'],
        'client_id': value['client_id'],
        'repository_generation': value['repository_generation'],
        'records': ((value['records'] as Array<any>).map(ObservationToJSON)),
    };
}